+++ helper.cpp
@@ -1,66 +1,71 @@
 #include "helper.h"
 
-SortCKKS::SortCKKS(std::string ccLocation, std::string pubKeyLocation, std::string multKeyLocation,
-                         std::string rotKeyLocation,
-                         std::string inputLocation,
-                         std::string outputLocation)
-    : m_PubKeyLocation(pubKeyLocation),
-      m_MultKeyLocation(multKeyLocation),
-      m_RotKeyLocation(rotKeyLocation),
-      m_CCLocation(ccLocation),
-      m_InputLocation(inputLocation),
-      m_OutputLocation(outputLocation)
-{
-    initCC();
-};
+// SortCKKS::SortCKKS(std::string ccLocation, std::string pubKeyLocation, std::string multKeyLocation,
+//                          std::string rotKeyLocation,
+//                          std::string inputLocation,
+//                          std::string outputLocation)
+//     : m_PubKeyLocation(pubKeyLocation),
+//       m_MultKeyLocation(multKeyLocation),
+//       m_RotKeyLocation(rotKeyLocation),
+//       m_CCLocation(ccLocation),
+//       m_InputLocation(inputLocation),
+//       m_OutputLocation(outputLocation)
+// {
+//     initCC();
+// };
 
-void SortCKKS::initCC(){
-    if (!Serial::DeserializeFromFile(m_CCLocation, m_cc, SerType::BINARY))
-    {
-        std::cerr << "Could not deserialize cryptocontext file" << std::endl;
-        std::exit(1);
-    }
+// void SortCKKS::initCC(){
+//     if (!Serial::DeserializeFromFile(m_CCLocation, m_cc, SerType::BINARY))
+//     {
+//         std::cerr << "Could not deserialize cryptocontext file" << std::endl;
+//         std::exit(1);
+//     }
 
-    if (!Serial::DeserializeFromFile(m_PubKeyLocation, m_PublicKey, SerType::BINARY))
-    {
-        std::cerr << "Could not deserialize public key file" << std::endl;
-        std::exit(1);
-    }
+//     if (!Serial::DeserializeFromFile(m_PubKeyLocation, m_PublicKey, SerType::BINARY))
+//     {
+//         std::cerr << "Could not deserialize public key file" << std::endl;
+//         std::exit(1);
+//     }
 
-    std::ifstream multKeyIStream(m_MultKeyLocation, std::ios::in | std::ios::binary);
-    if (!multKeyIStream.is_open())
-    {
-        std::exit(1);
-    }
-    if (!m_cc->DeserializeEvalMultKey(multKeyIStream, SerType::BINARY))
-    {
-        std::cerr << "Could not deserialize rot key file" << std::endl;
-        std::exit(1);
-    }
-    multKeyIStream.close();
+//     std::ifstream multKeyIStream(m_MultKeyLocation, std::ios::in | std::ios::binary);
+//     if (!multKeyIStream.is_open())
+//     {
+//         std::exit(1);
+//     }
+//     if (!m_cc->DeserializeEvalMultKey(multKeyIStream, SerType::BINARY))
+//     {
+//         std::cerr << "Could not deserialize rot key file" << std::endl;
+//         std::exit(1);
+//     }
+//     multKeyIStream.close();
 
-    std::ifstream rotKeyIStream(m_RotKeyLocation, std::ios::in | std::ios::binary);
-    if (!rotKeyIStream.is_open())
-    {
-        std::exit(1);
-    }
+//     std::ifstream rotKeyIStream(m_RotKeyLocation, std::ios::in | std::ios::binary);
+//     if (!rotKeyIStream.is_open())
+//     {
+//         std::exit(1);
+//     }
 
-    if (!m_cc->DeserializeEvalAutomorphismKey(rotKeyIStream, SerType::BINARY))
-    {
-        std::cerr << "Could not deserialize eval rot key file" << std::endl;
-        std::exit(1);
-    }
-    rotKeyIStream.close();
+//     if (!m_cc->DeserializeEvalAutomorphismKey(rotKeyIStream, SerType::BINARY))
+//     {
+//         std::cerr << "Could not deserialize eval rot key file" << std::endl;
+//         std::exit(1);
+//     }
+//     rotKeyIStream.close();
 
-    std::vector<Ciphertext<DCRTPoly>> deserializedCiphertexts;
-    std::ifstream in("in.bin", std::ios::binary);
-    if (!in.is_open()) {
-        std::cerr << "Error opening file for deserialization." << std::endl;
-        std::exit(1);
-    }
-    Serial::Deserialize(deserializedCiphertexts, in, SerType::BINARY);
-    in.close();
-}
+//     std::vector<Ciphertext<DCRTPoly>> deserializedCiphertexts;
+//     std::ifstream in("in.bin", std::ios::binary);
+//     if (!in.is_open()) {
+//         std::cerr << "Error opening file for deserialization." << std::endl;
+//         std::exit(1);
+//     }
+//     Serial::Deserialize(deserializedCiphertexts, in, SerType::BINARY);
+//     in.close();
+// }
+
+// void SortCKKS::eval(){
+//     std::cout << "To be filled" << std::endl;
+
+// }
 
 vector<Ciphertext<DCRTPoly>> encrypt_and_serialize_plaintext(
     CryptoContext<DCRTPoly> cc, KeyPair<DCRTPoly> keys, 
@@ -73,11 +78,14 @@ vector<Ciphertext<DCRTPoly>> encrypt_and_serialize_plaintext(
         plaintexts.push_back(p);
         ciphertexts.push_back(cc->Encrypt(keys.publicKey, p));
     }
-    // Serialize input
-    if (!Serial::SerializeToFile("in.bin", ciphertexts, SerType::BINARY)) {
+
+    // // Serialize input
+    if (!Serial::SerializeToFile(inputLocation, ciphertexts, SerType::BINARY)) {
         std::cerr << "Error serializing public key to public_key.txt." << std::endl;
         std::exit(1);
     }
+
+
     return ciphertexts;
 }
