+++ helper.h
@@ -22,24 +22,25 @@ static const string inputLocation = "./in.bin";
 static const string outputLocation = "./out.bin";
 
 
-class SortCKKS {
-    CryptoContext<DCRTPoly> m_cc;
-    PublicKey<DCRTPoly> m_PublicKey;
-    Ciphertext<DCRTPoly> m_InputC;
-    Ciphertext<DCRTPoly> m_OutputC;
-    string m_PubKeyLocation;
-    string m_MultKeyLocation;
-    string m_RotKeyLocation;
-    string m_CCLocation;
-    string m_InputLocation;
-    string m_OutputLocation;
+// class SortCKKS {
+//     CryptoContext<DCRTPoly> m_cc;
+//     PublicKey<DCRTPoly> m_PublicKey;
+//     Ciphertext<DCRTPoly> m_InputC;
+//     Ciphertext<DCRTPoly> m_OutputC;
+//     string m_PubKeyLocation;
+//     string m_MultKeyLocation;
+//     string m_RotKeyLocation;
+//     string m_CCLocation;
+//     string m_InputLocation;
+//     string m_OutputLocation;
 
-public:
-    SortCKKS(string ccLocation, string pubKeyLocation, string multKeyLocation,
-                string rotKeyLocation, string inputLocation, string outputLocation);
+// public:
+//     SortCKKS(string ccLocation, string pubKeyLocation, string multKeyLocation,
+//                 string rotKeyLocation, string inputLocation, string outputLocation);
 
-    void initCC();
-};
+//     void initCC();
+//     void eval();
+// };
 
 vector<Ciphertext<DCRTPoly>> encrypt_and_serialize_plaintext(CryptoContext<DCRTPoly> cc, KeyPair<DCRTPoly> keys, vector<Plaintext> plaintexts, vector<double> input);
 void serialize_keys(CryptoContext<DCRTPoly> cc);
