+++ sort_ckks.cpp
@@ -1,0 +1,78 @@
+#include "sort_ckks.h"
+
+SortCKKS::SortCKKS(std::string ccLocation, std::string pubKeyLocation, std::string multKeyLocation,
+                         std::string rotKeyLocation,
+                         std::string inputLocation,
+                         std::string outputLocation)
+    : m_PubKeyLocation(pubKeyLocation),
+      m_MultKeyLocation(multKeyLocation),
+      m_RotKeyLocation(rotKeyLocation),
+      m_CCLocation(ccLocation),
+      m_InputLocation(inputLocation),
+      m_OutputLocation(outputLocation)
+{
+    initCC();
+};
+
+void SortCKKS::initCC(){
+    if (!Serial::DeserializeFromFile(m_CCLocation, m_cc, SerType::BINARY))
+    {
+        std::cerr << "Could not deserialize cryptocontext file" << std::endl;
+        std::exit(1);
+    }
+
+    if (!Serial::DeserializeFromFile(m_PubKeyLocation, m_PublicKey, SerType::BINARY))
+    {
+        std::cerr << "Could not deserialize public key file" << std::endl;
+        std::exit(1);
+    }
+
+    std::ifstream multKeyIStream(m_MultKeyLocation, std::ios::in | std::ios::binary);
+    if (!multKeyIStream.is_open())
+    {
+        std::exit(1);
+    }
+    if (!m_cc->DeserializeEvalMultKey(multKeyIStream, SerType::BINARY))
+    {
+        std::cerr << "Could not deserialize rot key file" << std::endl;
+        std::exit(1);
+    }
+    multKeyIStream.close();
+
+    std::ifstream rotKeyIStream(m_RotKeyLocation, std::ios::in | std::ios::binary);
+    if (!rotKeyIStream.is_open())
+    {
+        std::exit(1);
+    }
+
+    if (!m_cc->DeserializeEvalAutomorphismKey(rotKeyIStream, SerType::BINARY))
+    {
+        std::cerr << "Could not deserialize eval rot key file" << std::endl;
+        std::exit(1);
+    }
+    rotKeyIStream.close();
+
+    if (!Serial::DeserializeFromFile(m_InputLocation, m_InputC, SerType::BINARY))
+    {
+        std::cerr << "Could not deserialize cryptocontext file" << std::endl;
+        std::exit(1);
+    }
+}
+
+void SortCKKS::eval(){
+
+    std::cout << "This is the sorting method that needs to be filled" << std::endl;
+    std::cout << "The output should be the ciphertext on m_OutputC" << std::endl;
+    
+    // To be filled
+
+}
+
+void SortCKKS::deserializeOutput(){
+
+    // if (!Serial::SerializeToFile(m_OutputLocation, m_OutputC, SerType::BINARY)) //To be fixed
+    if (!Serial::SerializeToFile(m_OutputLocation, m_InputC, SerType::BINARY))
+    {
+        std::cerr << " Could not serialize output ciphertext" << std::endl;
+    }
+}
